Move default doc options and fix imports
